requirements.txt
Flask==2.0.2
gunicorn==20.1.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.5.1
joblib>=1.2.0
Jinja2==3.0.2
Werkzeug==2.0.2
itsdangerous==2.0.1
click==8.0.1
python-dotenv==0.19.0
google-cloud-storage==2.5.0  # Added for GCS interaction
